% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_assertions.R
\name{assert_reactive}
\alias{assert_reactive}
\title{assert_reactive}
\usage{
assert_reactive(object, msg = "")
}
\arguments{
\item{object}{object to test}

\item{msg}{msg to be appended to the hard-coded default (string)}
}
\value{
invisible (TRUE) if expression is TRUE. Will error if is FALSE
}
\description{
assert_reactive
}
