% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_app_on_browser_exit.R
\name{close_app_on_browser_exit}
\alias{close_app_on_browser_exit}
\title{Stop App when browser closes}
\usage{
close_app_on_browser_exit()
}
\description{
Stops the app running if a user closes their web browser. Useful if distributing your app as a desktop / electron app.

Should be placed in the server function
}
\details{
WARNING: Not working ATM
}
