% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataframe.R
\name{download_dataframeServer}
\alias{download_dataframeServer}
\title{Title}
\usage{
download_dataframeServer(
  id,
  data_to_write,
  filename_full = "table.tsv",
  rownames = F,
  colnames = T
)
}
\arguments{
\item{optional_argument}{}
}
\value{

}
\description{
Title
}
