% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataframe.R
\name{download_dataframeUI}
\alias{download_dataframeUI}
\title{Download Dataframes}
\usage{
download_dataframeUI(
  id,
  label = "Download",
  tooltip_text = "",
  tooltip_pos = "right",
  style = "unite",
  color = "default",
  size = "sm"
)
}
\arguments{
\item{id}{shiny properties}
}
\value{

}
\description{
A shiny module that adds a download button to
}
