% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataframe.R
\name{mod_download_dataframe_server}
\alias{mod_download_dataframe_server}
\title{Download Dataframe}
\usage{
mod_download_dataframe_server(
  id,
  data_to_write,
  filename_full,
  rownames = FALSE,
  colnames = TRUE
)
}
\arguments{
\item{id}{Internal parameters for {shiny}}

\item{data_to_write}{dataframe or datatable (dataframe / data.table)}

\item{filename_full}{Name to save file under. Include any extensions (string)}

\item{rownames}{Save row names? (boolean)}

\item{colnames}{Save column names? (boolean)}
}
\value{
None, Run for its side effects. Downloads a dataframe/data.table to the local machine.
}
\description{
Downloads a dataframe or datatable.
}
