% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataframe.R
\name{mod_download_dataframe_ui}
\alias{mod_download_dataframe_ui}
\title{Download Dataframes}
\usage{
mod_download_dataframe_ui(
  id,
  label = "Download",
  tooltip_text = "",
  tooltip_pos = "right",
  style = "unite",
  color = "default",
  size = "sm"
)
}
\arguments{
\item{id}{Internal parameters for {shiny}}

\item{label}{The label that should appear on the button.}

\item{tooltip_text}{text to appear when cursor hovers over download button (string)}

\item{tooltip_pos}{Where the tooltip should appear relative to its target (top, bottom, left, or right). Defaults to "right"}

\item{style}{Style of the button, to choose between \code{simple}, \code{bordered},
\code{minimal}, \code{stretch}, \code{jelly}, \code{gradient}, \code{fill},
\code{material-circle}, \code{material-flat}, \code{pill}, \code{float}, \code{unite}.}

\item{color}{Color of the button : \code{default}, \code{primary}, \code{warning},
\code{danger}, \code{success}, \code{royal}.}

\item{size}{Size of the button : \code{xs},\code{sm}, \code{md}, \code{lg}.}
}
\description{
A shiny module that adds a download button to
}
