% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataframe.R
\name{mod_download_dataframeServer}
\alias{mod_download_dataframeServer}
\title{Download Dataframe}
\usage{
mod_download_dataframeServer(
  id,
  data_to_write,
  filename_full = "table.tsv",
  rownames = FALSE,
  colnames = TRUE
)
}
\arguments{
\item{id}{Internal parameters for {shiny}}

\item{data_to_write}{dataframe or datatable (dataframe / data.table)}

\item{filename_full}{"filepath to save file as" (string)}

\item{rownames}{"Save Rownames (boolean)"}

\item{colnames}{""}
}
\value{
None, Run for its side effects.
}
\description{
Should also work for datatables. May want to create another function that uses data.table to do this faster
}
